Fix bug where three bib fields saved wrong index
This patches a minor issue where three rules were calling builder.add
with the wrong - or a missing - index argument. This does not prevent
the statement from being saved, but saves it with a missing/invalid
index column, which will prevent multiple values from being saved on the
same predicate as well as potentially prevent source_record_path from
being saved with the record. The effected predicates are Genre/form,
Part of, and Call Number