Add serialization of Contents, Contents title
Adds extraction of new literals: 'Contents', 'Contents title'. Includes
tests against 1) basic 505 blocks (ind2 ' '), which mash all content in $a
and 2) "enhanced" 505s (ind2 '0'), which divide content across $g, $r,
and $t. Ordering of multiple matches is expected to be important, but
we're relying on implicit ordering in store (which implicitly controls
ordering in final jsonld response, which in turns controls front-end
display).

https://jira.nypl.org/browse/SCC-469