Fixes improper handling of malformed 880$6
When parallel fields exist for a given var block, they rely on the following convention:
 - $6 has a value like "880-01" (or, generally, "880-N" for N >= 1)
 - For each such link, there will be a 880 block with a $6 value like "264-03/(2/r"

The link *to* the 880 block and the link *back* from the 880 block are
necessary to ensure the parallel value in the 880 block pairs correctly
with the primary value.

If, for a given forward parallel link, the back-link is malformed (e.g.
"264 -03/(2/r"), we should intentionally fail the parallel value.

This PR patches the existing behavior, which breaks extraction entirely
when the back-link is malformed. I've added a test confirming the fix
against a sample record that recently gummed up the pipeline.