SCC-2206 Add Holding class to discovery-store-poster
This creates a set of new classes and functions for handling Holding records coming from the new Holding pipeline. These represent serial holdings that are to be represented in the SCC front-end.

These records function very similarly to Item records, in that they are related to Bibs in a (generally) one-to-many relationship. The records themselves are fairly simple and the a new mapping is being created in NYPL-core to represent them.

In addition to adding the new Holding updater this adds some new logic to the index to properly route all new records.

No new tests have been created yet, these will be added soon, but wanted to validate this general approach to extending the store-poster before delving into those issues