NOREF Remove ProcessViz class to eliminate dependencies
This removes the `ProcessViz` class as it has a dependency that is no longer compatible with `node6.x` (specifically `chalk` now uses a spread operator). This class was used to display a nice graphical output for ingest jobs, and can be removed because these jobs are now rarely, if ever, run and if/when they are the output does not need to be "fancy". The ingest process will still operate in the same way, simply with less information presented to the user.

While the `ProcessViz` class being removed is the general solution, the specific fix is the removal of the `blessed-contrib` library, which is a dependency of that class. To keep things neat unused classes have been removed from `package.json` and a deprecation comment added to the README.