NOREF Update function to nodejs10.x and async execution
This is a refactor of the function to bring it in line with nodejs10.x and update the handler to be asynchronus, which hopefully has a performance boost.

This function was bumped to nodejs10.x previously for compatibility with AWS but none of the code was updated. This performs housekeeping to upgrade packages (removing a number of unused modules), lint the function with new standards and update the test suite.

This also changes the manner in which the `handler` is invoked. Previously (as noted in a comment in the code) this relied on the event loop and a callback to single that execution was complete. But this seems to cause a condition where the function hangs for 30 seconds after completion. While it is unclear exactly why this is, hopefully switching this out for a Promise will eliminate this wait and result in a more timely processing speed for this function.

Due to the necessity of linting a large number of files for nodejs10.x this is a rather noisy commit/PR, any significant changes will be highlighted with comments.