HOTFIX Missing bib_ids for items and holdings
This is a fix for a bug that has appeared that affects the lookup of `bib_ids` for `Item` and `Holding` records. It is still unclear what the root source of this is but it seems that there can be records that are pushed to the `discovery-store` for the first time without a `bib_id`, either due to a cataloging error or because the record was deleted already(?)

If this is true there is no need to re-index any bib records because as far as the store and api are concerned, no updates have taken place. The record should already not exist and can be safely skipped. I've added a warning so that these events are logged.

The only question is if there is a more fundamental reason for this issue that should be addressed at that level