Improves handling of missing item type
Currently, if an item has a Item Type that is not mapped in NYPL Core,
the poster errors hard. This fixes that by skipping the Item Type check
for unmapped Item Types. When determining whether or not an item is
"research", Item Type is one of multiple checks we make. If an item has
an unmapped Item Type, this change will cause it's Item Type to fail to
be contribute to its designation as a "research" item; It may still be
considered "research" if other checks pass.

This fix will accompany an update to NYPL-Core to add missing Item Types