Ensure ItemType always included, suppress otf
This fixes an issue with how suppressed items are handled. Prior to this
change, the item serializer returned early when encountering a
suppressed item without extracting any of the typical fields, including
catalogItemType. CatalogItemType is essential for determinining whether
or not an item is Research downstream. Knowing whether or not an item is
Research even when it's suppressed has implications on the is-research
check for bibs. This fix ensures that the set of statements returned for
suppressed items includes catalogItemType by simply moving the code that
generates the catalogItemType statement into position *before* the
suppression check.

This change is being made mainly to address a bug where bibs with solely
suppressed items are mis-classified as non-research because their items
aren't identified as research (due to not have identified
catalogItemTypes).

Additionally, this update includes a fix to how we handle OTF records,
which have heretofore been accidentally (but correctly) hidden by the
bug that this update fixes. (OTF records consist of a non-suppressed bib
and an item with Item Type 50, which this app specially considers
suppressed, so OTF records have been suppressed due to the above
described bug.) This update includes code to identify OTF records by
location code 'os' and mark them suppressed.

https://jira.nypl.org/browse/SCC-2517