Fix parallel extraction
Changes how parallel fields are looked up to use a new .isParallelFor
property of field mappings, which allows us to define marc paths once in
the primary field and use them when extracting parallel values. This
means 1) we are guaranteed to use the same series of marc paths and in
the same order when extracting parallels as we do when extracting
primary values, and 2) we can ensure that the index of a primary value
corresponds with the index of a parallel value. Also ensures that
blank statements are not used except when necessary to pad values (to
protect index relationships between primary and parallel values). Also,
"blank" values are now always '' instead of ' '.

https://jira.nypl.org/browse/SCC-3071